requirements.txt
requests
beautifulsoup4
pillow
numpy
playwright
playwright-stealth
